requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py36_0
astor=0.8.1=py36_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.7.22=0
cairo=1.14.12=h8948797_3
certifi=2020.6.20=py36_0
cudatoolkit=8.0=3
cudnn=7.1.3=cuda8.0_0
cupti=8.0.61=0
easydict=1.9=pypi_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.2=h5ab3b9f_0
gast=0.4.0=py_0
glib=2.65.0=h3eb4bd4_0
graphite2=1.3.14=h23475e2_0
grpcio=1.31.0=py36hf8bcb03_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
importlib-metadata=1.7.0=py36_0
intel-openmp=2020.2=254
jasper=2.0.14=h07fcdf6_1
jpeg=9b=h024ee3a_2
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.12.4=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_1
markdown=3.2.2=py36_0
mkl=2020.2=256
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.1.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
neuralgym=0.0.1=pypi_0
numpy=1.16.4=pypi_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1g=h7b6447c_0
pcre=8.44=he6710b0_0
pillow=7.2.0=pypi_0
pip=20.2.2=py36_0
pixman=0.40.0=h7b6447c_0
progressbar=2.5=pypi_0
protobuf=3.12.4=py36he6710b0_0
py-opencv=3.4.2=py36hb342d67_1
python=3.6.12=hcff3b4d_2
pyyaml=5.3.1=pypi_0
readline=8.0=h7b6447c_0
scipy=0.19.0=pypi_0
setuptools=49.6.0=py36_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tensorboard=1.10.0=py36hf484d3e_0
tensorflow=1.10.0=gpu_py36h8dbd23f_0
tensorflow-base=1.10.0=gpu_py36h3435052_0
tensorflow-gpu=1.10.0=hf154084_0
termcolor=1.1.0=py36_1
tk=8.6.10=hbc83047_0
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
